Extend `for` loop with support for expanding `Iterable`s.
For example, if you have bound the following json array to `jsonArray`:

```
[
  ["a", "b", "c"],
  ["d", "e", "f"],
  ["g", "h", "i"]
]
```

The following template

```
{% for x, y, z in jsonArray %}{{ z }}/{{ y }}/{{ x }}
{% end %}
```

renders to:

```
c/b/a
f/e/d
i/h/g
```

This commit also extends `MapBindings` and `EntryBindings` to be `Iterable` which allows you to do this:

```
{% for key, value in $json %}
{{ key }} -> {{ value }}
{% end %}
```

Note that expansion is rather tolerant and it's not required that the number of variables and values match.
The only error condition is when you actually access an "unbound" variable (i.e. a variable which has no corresponding value).

If only one loop identifier is given, the current behavior doesn't change.