Extend `SetTag` with support for non-block assignments.
Examples

Assign a single value:

```
  {% set foo = "bar" %}
```

Unpack multiple values (which requires the expression to the right to evaluate to an `Iterable`).

```
  {% set foo, bar = foobar %}
```

Note, if the right side evaluates to an `Iterable` but there is only one variable to the left, the whole `Iterable` will be assigned to the variable, not only the first value of it.