Next stage in refactoring expression parsing.
This commit introduces "Iterations" which is just a comma separated list of terms. Function parameters are parsed as iterations now too.

You can use them in various places, e.g.

**Set multiple values in a single `set` statement**
```
{% set foo, bar = "bing", "bong" %}
```
This sets `foo="bing"` and `bar="bong"`.

**Swap two (or more) values**
```
{% set foo, bar = bar, foo %}
{% set foo, bar, baz = bar, baz, foo %}
```

**Iterate over multiple values**
```
{% for a in 1, 2, 3 %} -{{ a }}- {% end %}
```
This results in ` -1-  -2-  -3- `

**Iterate over multiple iterables**
```
{% for number, letter in (1, "a"), (2, "b"), (3, "c") %} {{ number }}:{{ letter }} {% end %}
```
This results in ` 1:a  2:b  3:c `

Iterations behave pretty much like tuples in Python, i.e. `a` or `(a)` is a scalar (not iterable unless `a` iteself is iterable), `a,` or `(a,)` is an Iterable with a single value `a`, `(a,b)` is an iterable with two values `a` and `b`, etc

Note, this commit adds a library of ours which we use for the `Flattened` and `SingletonIterator` iterators.

TODO: refactor StrictIterationTermParser and LaxIterationTermParser. Both share a lot of code.