Add support for Python-style `not`, `and` and `or` operators as alternatives to the C-style `!`, `&&` and `||`.
Also add support for the binary `in` operator which tests if a value exists in an iterable.

Note, this only works if value supports the `equals()` contract, which means it can not be an iteration.

So this works
```
{% if "a" in ("a", "b", "c") %}…
```

This doesn't work (yet):
```
{% if ("a", 1) in (("a", 1), ("b", 2), ("c", 2)) %}…
```

Possible extension:

* If in `a in b` `a` is a `char` or a `CharSequence` and `b` is a `CharSequence` return `true` if `b` contains `a` (just like in Python, where `"el" in "hello"` is `True`).
* Support iterations

General note regarding iterations: Although mostly not required, it's sometimes necessary to enclose them in parenthesis to enfore the correct evaluation order. `,` has the lowest precendece hence `a in 1, 2, 3` would be intepreted as `(a in 1), (2), (3)`, which is a non-empty iterable (and always evaluated to `true`).  The correct syntax is `a in (1, 2, 3)`.