Add support for Python-style `not`, `and` and `or` operators as alternatives to the C-style `!`, `&&` and `||`.
Also add support for the binary `in` operator.

General note regarding iterations: Although mostly not required, it's sometimes necessary to enclose them in parenthesis to enfore the correct evaluation order. `,` has the lowest precendece hence `a in 1, 2, 3` would be intepreted as `(a in 1), (2), (3)`, which is a non-empty iterable (and always evaluated to `true`).  The correct syntax is `a in (1, 2, 3)`.