Replace `Function` and `Variable` with `AccessTerm`s.
An `AccessTerm` is a binary term with an access operator. There are three forms:

1. `a.b` in which `b` identifies a key, field or method of object `a`.
2. `a[b]` which is like 1. but also allows `b` to be numeric or an expression
3. `a(b)` which invokes the method `a` with the parameters `b` with `b` being an "iteration".

These three forms can be combined in any way, e.g.
```
a.b["key"].method(param1, param2).field
```

Note, higher order functions are not supported (which means a method can not return a function), so this is **not** possible:

```
a.b(x)(y)
```

Most of the "access" logic is just copied from the old classes.

This commit also adds indexed access to iterations so you can do this:

```
{% set iteration = "a", "b", "c" %}
{{ iteration[1] }}
```
which echos `b` (the index is 0-based).