Fix associativity of binary terms.
Operators with the same precedence are usually evaluated left to right (with very few exceptions), see https://docs.oracle.com/javase/tutorial/java/nutsandbolts/operators.html

For instance the term `8 - 4 - 2` is commonly understood to be equivalent to `(8 - 4) - 2 == 2`.

However, before this commit the term had been parsed into the following tree:

```
      (-)
     /   \
    8    (-)
        /   \
       4     2
```

Which had been evaluated to `8 - (4 - 2) == 6`

With this commit it's parsed correctly into this tree:

```
      (-)
     /   \
   (-)    2
  /   \
 8     4
```

Which is evaluated to `(8 - 4) - 2 == 2`.