Fixes #37 by adding the ability to escape Java identifiers with \.
Shouldn't normally be needed, as for the most part you'll control the
input, but if you need to you can now escape Java identifier with "\",
so the following will work:

{{ my\-var }}

Will print the value of a variable called "my-var".

Not documenting this, as I don't really want to encourage it.